requirements.txt
# Vercel deployment - Absolute minimum to stay under 250MB
# For Render.com and other platforms: use requirements-dev.txt instead
# WARNING: This may not support all features. For full functionality, run locally.

flask==3.0.0
flask-login==0.6.3
flask-mail==0.9.1
werkzeug==3.0.1
bcrypt==4.1.2
itsdangerous==2.1.2
python-dotenv==1.0.0
gunicorn==21.2.0

# Absolute minimum data processing stack
# Using oldest stable versions to minimize size
pandas==1.3.5
numpy==1.21.6
openpyxl==3.0.9

# These are small and required
python-dateutil==2.8.2
requests==2.31.0

# Estimated total: ~170-190MB (should be under 250MB)
